fastapi_notifications/src/requirements.txt
fastapi==0.111.0
gunicorn==23.0.0
pydantic-settings==2.7.1

celery==5.4.0

redis==5.0.4

pika==1.3.2
aio-pika==9.5.4

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
werkzeug==3.1.3

alembic==1.14.0
typer==0.15.1

psycopg==3.2.3
psycopg-binary==3.2.3

httpx[http2]==0.28.1
httpx-oauth==0.16.1

python-jose==3.3.0
aiosmtplib==4.0.0